refactor: change TransakConnectionPanel to be a popup banner
