+++ packages/desktop/views/dashboard/buy-sell/components/TransakConnectionBanner.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { TRANSAK_WIDGET_URL } from '@auxiliary/transak'
     import { TransakConnectionStatus } from '@auxiliary/transak/enums'
-    import { IconButton, IconName, Pill, Text, TooltipIcon, Tooltip } from '@bloomwalletio/ui'
+    import { IconButton, IconName, Text, TooltipIcon, Tooltip } from '@bloomwalletio/ui'
     import { Platform } from '@core/app'
     import { localize } from '@core/i18n'
     import { tick } from 'svelte'
@@ -40,22 +40,11 @@
     }
 </script>
 
-<div class="flex flex-col items-center w-full p-6 gap-4">
-    <div class="w-full flex justify-between items-center h-8">
-        <img data-label="transak-logo" width="90" height="28" src="assets/logos/transak.svg" alt="Transak" />
-        <div>
-            {#if connectionStatus === TransakConnectionStatus.Connected}
-                <Pill color="success">{localize('general.connected')}</Pill>
-            {:else if connectionStatus === TransakConnectionStatus.Redirected}
-                <Pill color="warning">{localize('general.redirected')}</Pill>
-            {:else if connectionStatus === TransakConnectionStatus.Waiting}
-                <Pill color="neutral">{localize('general.waitingConnection')}</Pill>
-            {:else}
-                <Pill color="danger">{localize('general.disconnected')}</Pill>
-            {/if}
-        </div>
-    </div>
-    <div class="flex flex-row justify-between w-full items-center">
+<div class="flex flex-row items-center w-full p-6 gap-4 bg-surface-1 dark:bg-surface-1-dark rounded-t-[32px]">
+    <img class="w-10 h-10" src="assets/logos/transak-icon.svg" alt="Transak" />
+
+    <div class="flex flex-col items-start gap-2">
+        <img class="h-4" src="assets/logos/transak-text.svg" alt="Transak" />
         <div class="flex items-center gap-1 w-full h-3">
             {#if connectionStatus === TransakConnectionStatus.Connected}
                 <TooltipIcon
@@ -93,11 +82,13 @@
                 <Tooltip anchor={textContainer} event="hover" placement="top" text={url ?? TRANSAK_WIDGET_URL} />
             {/if}
         </div>
+    </div>
+    <div class="absolute top-8 right-16 flex flex-row justify-between items-center">
         <Tooltip anchor={refreshButton} text={localize('actions.refresh')} event="hover" placement="top" />
         <div bind:this={refreshButton}>
             <IconButton
                 icon={IconName.Refresh}
-                size="xs"
+                size="sm"
                 on:click={refreshFunction}
                 disabled={connectionStatus === TransakConnectionStatus.Waiting}
             />
