+++ packages/desktop/views/dashboard/buy-sell/components/TransakExchangePanel.svelte
@@ -161,7 +161,6 @@
                 paymentMethod: selectedPaymentOption?.value ?? '',
                 recipientAddress: selectedRecipient?.address ?? '',
             },
-            hideClose: true,
             confirmClickOutside: true,
         })
     }
